TMBF: Complete conversion of EPICS waveform scaling
Now all the main EPICS waveforms are scaled to eliminate arbitrary
units.  Compensation filters take values in the range [-2,2), and DAC gain,
FIR filteres, and detector window are all in the range [-1,1).

Note that buffer and DDR readouts are still in raw integer units.

git-svn-id: svn+ssh://serv2/home/subversion/repos/controls/diamond/trunk/ioc/TMBF@115010 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ tmbfApp/src/hardware.c
@@ -425,7 +425,7 @@ void hw_write_adc_offsets(short offsets[4])
         config_space->adc_offsets = offsets[i];
 }
 
-void hw_write_adc_filter(short taps[12])
+void hw_write_adc_filter(int taps[12])
 {
     config_space->write_select = 0;
     for (int i = 2; i >= 0; i --)
@@ -494,7 +494,7 @@ void hw_write_dac_enable(bool enable)
     WRITE_CONTROL_BITS(0, 1, enable);
 }
 
-void hw_write_dac_preemph(short taps[3])
+void hw_write_dac_preemph(int taps[3])
 {
     config_space->write_select = 0;
     for (int i = 2; i >= 0; i --)
@@ -726,7 +726,7 @@ void hw_write_det_gain(unsigned int gain)
     WRITE_CONTROL_BITS(24, 3, gain);
 }
 
-void hw_write_det_window(const uint16_t window[DET_WINDOW_LENGTH])
+void hw_write_det_window(const int window[DET_WINDOW_LENGTH])
 {
     LOCK();
     config_space->write_select = 1;    // Select sequencer window
